fix midi (T.Yui)
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@64 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ win9x/cmserial.cpp
@@ -53,7 +53,7 @@ static BYTE serialgetstat(COMMNG self) {
 	}
 }
 
-static UINT serialmsg(COMMNG self, UINT msg, long param) {
+static long serialmsg(COMMNG self, UINT msg, long param) {
 
 	(void)self;
 	(void)msg;
